quoteScriptAgent/requirements.txt
fastapi
uvicorn
langchain
chromadb
openai
python-dotenv
langchain-openai
